Fixed WEM plugin for Scope.
WEMs can be played as long as the plugin can convert them.
Move RelayCommand to Utils project.

+++ src/Plugins/Scope.FileViewer.WEM/Models/WemFile.cs
@@ -6,21 +6,19 @@ using WEMSharp;
 
 namespace Scope.FileViewer.WEM.Models
 {
-  internal class WemFileViewer : IFileViewer
+  internal class WemFile : IFileViewer
   {
     private readonly IFile _file;
+    private readonly IMessageQueue _messageQueue;
     private byte[] _rawData;
 
-    public WemFileViewer(IFile file)
+    public WemFile(IFile file, IMessageQueue messageQueue)
     {
       _file = file;
+      _messageQueue = messageQueue;
     }
 
-    public string FileName => _file.Name;
-    //public string PathInternal => _file.FullPath;
-    //public DateTime LastModified => _file.LastModifiedDate;
-
-    public string Header { get; }
+    public string Header => _file.Name;
 
     public async Task<int> GetNumberOfRawBytesAsync()
     {
@@ -43,21 +41,48 @@ namespace Scope.FileViewer.WEM.Models
 
     public async Task<string> ConvertAsync()
     {
+      if (_rawData == null)
+      {
+        _rawData = await Task.Run(() => GetRawBytes());
+      }
+
       var tmpDirectory = Path.GetTempPath();
 
-      var fileName = $"Crucible_tmpWEM_{Guid.NewGuid()}";
+      var fileName = $"Scope_tmpWEM_{_file.Path.Replace('/','.')}";
       var wemFileName = $"{fileName}.wem";
       var oggFileName = $"{fileName}.ogg";
 
       var wemFilePath = Path.Combine(tmpDirectory, wemFileName);
       var oggFilePath = Path.Combine(tmpDirectory, oggFileName);
 
-      File.WriteAllBytes(wemFilePath, _rawData);
+      if (File.Exists(wemFilePath))
+      {
+        _messageQueue.Add($"Found {wemFilePath}");
+      }
+      else
+      {
+        File.WriteAllBytes(wemFilePath, _rawData);
+        _messageQueue.Add($"Wrote {wemFilePath}");
+      }
 
-      var wemFile = new WEMFile(wemFilePath, WEMForcePacketFormat.NoForcePacketFormat);
-      //MainWindow.SetStatus($"Wrote {wemFilePath}");
+      if (File.Exists(oggFilePath))
+      {
+        _messageQueue.Add($"Found {wemFilePath}");
+        return oggFilePath;
+      }
+
+      WEMFile wemFile;
+      try
+      {
+        wemFile = new WEMFile(wemFilePath, WEMForcePacketFormat.NoForcePacketFormat);
+      }
+      catch (Exception)
+      {
+        throw;
+      }
 
       var codebookPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, @"Plugins\Resources\packed_codebooks_aoTuV_603.bin");
+
       try
       {
         wemFile.GenerateOGG(oggFilePath, codebookPath, false, false);
@@ -72,14 +97,13 @@ namespace Scope.FileViewer.WEM.Models
 
       var exitCode = await External.RunProcessAsync(revorbPath, oggFilePath);
 
-      //MainWindow.SetStatus($"Wrote {oggFilePath}");
+      _messageQueue.Add($"Wrote {oggFilePath}");
 
       return oggFilePath;
     }
 
     public void Dispose()
-    {
-      
+    {      
     }
   }
 }
